kv - changed things
