+++ frontend/src/main/components/Staff/StaffTable.js
@@ -1,33 +1,33 @@
-import React from "react";
- import OurTable from "main/components/OurTable"
-//  import { useBackendMutation } from "main/utils/useBackend";
-//  import { cellToAxiosParamsDelete, onDeleteSuccess } from "main/components/Utils/CoursesUtils"
-//  import { useNavigate } from "react-router-dom";
-//  import { hasRole } from "main/utils/currentUser";
+// import React from "react";
+//  import OurTable from "main/components/OurTable"
+// //  import { useBackendMutation } from "main/utils/useBackend";
+// //  import { cellToAxiosParamsDelete, onDeleteSuccess } from "main/components/Utils/CoursesUtils"
+// //  import { useNavigate } from "react-router-dom";
+// //  import { hasRole } from "main/utils/currentUser";
 
- export default function StaffTable({ staff }) {
+//  export default function StaffTable({ staff }) {
 
-    // const navigate = useNavigate();
-
-
-     const columns = [
-         {
-             Header: 'id',
-             accessor: 'id',
-         },
-         {
-             Header: 'courseId',
-             accessor: 'courseId',
-         },
-         {
-             Header: 'userId',
-             accessor: 'userId',
-         }
-     ];
+//     // const navigate = useNavigate();
 
 
-     return <OurTable
-         data={staff}
-         columns={columns}
-         testid={"StaffTable"} />;
-    };
+//      const columns = [
+//          {
+//              Header: 'id',
+//              accessor: 'id',
+//          },
+//          {
+//              Header: 'courseId',
+//              accessor: 'courseId',
+//          },
+//          {
+//              Header: 'userId',
+//              accessor: 'userId',
+//          }
+//      ];
+
+
+//      return <OurTable
+//          data={staff}
+//          columns={columns}
+//          testid={"StaffTable"} />;
+//     };
